Fix Win64 build after r16079.
TBR=yangguo@chromium.org

Review URL: https://codereview.chromium.org/22473002

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@16083 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/runtime.cc
@@ -8547,7 +8547,8 @@ RUNTIME_FUNCTION(MaybeObject*, Runtime_CompileForOnStackReplacement) {
 
     // Use linear search of the unoptimized code's back edge table to find
     // the AST id matching the PC.
-    uint32_t target_pc_offset = frame->pc() - unoptimized->instruction_start();
+    uint32_t target_pc_offset =
+      static_cast<uint32_t>(frame->pc() - unoptimized->instruction_start());
     uint32_t loop_depth = 0;
 
     for (FullCodeGenerator::BackEdgeTableIterator back_edges(*unoptimized);
